avance gasto persona

+++ src/flux/expenseActions.js
@@ -1,6 +1,6 @@
 // En tu archivo taskActions.js
 import axios from 'axios';
-import { assignGastoPersona } from './personExpenseActions.js';  // Importa la acción de personExpenseActions
+import { assignGastoPersona, deleteGastoPersonaByGasto } from './personExpenseActions.js';  // Importa la acción de personExpenseActions
 
 // tipos de acciones
 export const ADD_EXPENSE = 'ADD_EXPENSE';
@@ -68,7 +68,7 @@ export const getExpensesByUnit = (unitId) => {
 
                 if (response.status === 200) {
                     // Guardar los gastos en el estado global antes de devolver la respuesta
-                    dispatch(saveFetchedExpensesData(response.data));
+                    dispatch(addExpense(response.data));
                     // Resolver la Promesa con los datos
                     resolve(response.data);
                 } else {
@@ -97,33 +97,50 @@ export const deleteExpense = (expenseId) => {
 export const deleteExpenseFromDatabase = (expenseId) => {
     return async (dispatch) => {
         try {
-            // Realiza la solicitud DELETE al endpoint para eliminar el gasto por su ID
+            // Llama a la acción para obtener los datos actualizados de gastos
+            const updatedExpenses = await dispatch(getExpensesByUnit(localStorage.getItem('id_unidad')));
+
+            // Busca el gasto correspondiente en los datos actualizados
+            const targetExpense = updatedExpenses.gastos.find(expense => expense.id === expenseId);
+
+            if (!targetExpense) {
+                console.error('Gasto no encontrado en los datos actualizados.');
+                return null;
+            }
+
+            // Llama a la acción para obtener los detalles del gasto usando la factura
+            await dispatch(getGastoDetails(targetExpense.factura));
+
+            // Llama a la acción para eliminar gastos persona por el ID del gasto asociado
+            await dispatch(deleteGastoPersonaByGasto(targetExpense.factura));
+
+            // Obtiene y guarda los datos actualizados de gastos utilizando la acción correspondiente
+            const updatedExpensesAfterDelete = await dispatch(getExpensesByUnit(localStorage.getItem('id_unidad')));
+
+            // Despacha la acción para eliminar el gasto del estado global después de obtener los datos actualizados
+            dispatch(deleteExpense(expenseId));
+
+            // Realiza la solicitud DELETE al endpoint para eliminar el gasto por su ID al final
             const response = await axios.delete(`http://localhost:5000/delete_gasto_por_unidad/${expenseId}`);
 
-            if (response.status === 200) {
-                // Despacha la acción para eliminar el gasto del estado global
-                dispatch(deleteExpense(expenseId));
-
-                // Obtiene y guarda los datos actualizados de gastos utilizando la acción correspondiente
-                const updatedExpenses = await dispatch(getExpensesByUnit(localStorage.getItem('id_unidad')));
-                dispatch(saveFetchedExpensesData(updatedExpenses));
-
-                // Devuelve los datos actualizados de gastos después de la eliminación
-                return updatedExpenses;
-            } else {
+            if (response.status !== 200) {
                 // Maneja el caso en que la respuesta del servidor no sea exitosa
                 console.error('Error al eliminar el gasto:', response.data.error);
                 // Devuelve null o algún valor que indique que la eliminación falló
                 return null;
             }
+
+            // Devuelve los datos actualizados de gastos después de la eliminación
+            return updatedExpensesAfterDelete;
         } catch (error) {
-            // Muestra un mensaje de error si ocurre un error durante la eliminación del gasto
+            // Maneja los errores según sea necesario
             console.error('Error durante la eliminación del gasto:', error);
             // Devuelve null o algún valor que indique que la eliminación falló
             return null;
         }
     };
 };
+
 
 // Acción para obtener detalles del gasto
 export const getGastoDetails = (factura) => async (dispatch) => {
@@ -133,18 +150,18 @@ export const getGastoDetails = (factura) => async (dispatch) => {
         if (response.status === 200) {
             const gastoDetails = response.data.gasto;
 
-            // Asegúrate de extraer correctamente el id del objeto gastoDetails
+            // Asegúrate de extraer correctamente el ID y el número de factura del objeto gastoDetails
             const gastoId = gastoDetails ? gastoDetails.id : null;
+            const gastoFactura = gastoDetails ? gastoDetails.factura : null;
 
-            if (gastoId) {
+            if (gastoId && gastoFactura) {
                 console.log('Detalles del gasto obtenidos con éxito:', gastoDetails);
                 dispatch({ type: GET_GASTO_DETAILS_SUCCESS, payload: gastoDetails });
 
                 // Agrega un console.log para la data que se está enviando al estado
                 console.log('Data enviada al estado:', gastoDetails);
-
             } else {
-                console.error('Error: No se pudo obtener el id del gasto desde los detalles:', gastoDetails);
+                console.error('Error: No se pudo obtener el ID o el número de factura del gasto desde los detalles:', gastoDetails);
                 dispatch({ type: GET_GASTO_DETAILS_ERROR, payload: 'Error al obtener detalles del gasto' });
             }
 
@@ -159,6 +176,7 @@ export const getGastoDetails = (factura) => async (dispatch) => {
         dispatch({ type: GET_GASTO_DETAILS_ERROR, payload: 'Error al obtener detalles del gasto' });
     }
 };
+
